+++ src/redturtle/volto/restapi/services/search/get.py
@@ -16,9 +16,12 @@ class SearchHandler(OriginalHandler):
         else:
             fullobjects = False
 
+        if "use_site_search_settings" in query:
+            del query["use_site_search_settings"]
+            query = self.filter_query(query)
+
         self._constrain_query_by_path(query)
         query = self._parse_query(query)
-
         # se non passiamo i parametri con **, non viene controllato bene
         # il permesso di vedere i contenuti scaduti quando si passa un path
         lazy_resultset = self.catalog.searchResults(**query)